InfoMe/requirements.txt
decorator==4.1.2
Django==1.11.4
django-extensions==1.9.0
djangorestframework==3.6.4
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
python-apt==1.1.0b1
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
unattended-upgrades==0.1
wcwidth==0.1.7
